requirements.txt
asgiref==3.11.1
dj-database-url==3.1.2
Django==5.2.12
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
numpy==2.2.6
pandas==2.3.3
psycopg2-binary==2.9.11
PyJWT==2.12.1
python-dateutil==2.9.0.post0
pytz==2026.1.post1
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
python-dotenv==1.0.1
